added token verification for delete request

+++ controllers/blogs.js
@@ -39,10 +39,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const blog = await Blog
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (user._id.toString() !== blog.user._id.toString()) {
+        return response.status(401).json({ error: 'invalid token' })
+    }
+
+    const blogDelete = await Blog
         .findByIdAndRemove(request.params.id)
 
-    response.status(204).json(blog)
+    response.status(204).json(blogDelete)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
